actualizacion de MVC

+++ app/src/main/java/com/fpl/mantenimientovehicular/model/ModeloVehiculo.java
@@ -17,7 +17,8 @@ public class ModeloVehiculo {
     private String anho;
     private String placa;
     private String tipo;
-    private int kilometraje;
+    private int kilometraje_actual;
+    private int kilometraje_esperado;
     private static DataBaseHelper dbHelper;
     private static SQLiteDatabase db;
     private String table = "Vehiculo";
@@ -27,13 +28,14 @@ public class ModeloVehiculo {
         dbHelper = new DataBaseHelper(context);
         db = dbHelper.getWritableDatabase();
     }
-    public ModeloVehiculo(int id, String marca, String anho, String placa, String tipo, int kilometraje) {
+    public ModeloVehiculo(int id, String marca, String anho, String placa, String tipo, int kilometraje_actual, int kilometraje_esperado) {
         this.id = id;
         this.marca = marca;
         this.anho = anho;
         this.placa = placa;
         this.tipo = tipo;
-        this.kilometraje = kilometraje;
+        this.kilometraje_actual = kilometraje_actual;
+        this.kilometraje_esperado = kilometraje_esperado;
     }
     public void open() {
         db = dbHelper.getWritableDatabase();
@@ -49,12 +51,16 @@ public class ModeloVehiculo {
     public long agregar(ModeloVehiculo vehiculo) {
 //        try(SQLiteDatabase db = dbHelper.getWritableDatabase()) {
         try{
+            if (db == null) {
+                throw new Exception("Error al abrir la base de datos");
+            }
             ContentValues values = new ContentValues();
             values.put("marca", vehiculo.getMarca());
             values.put("anho", vehiculo.getAño());
             values.put("placa", vehiculo.getPlaca());
             values.put("tipo", vehiculo.getTipo());
-            values.put("kilometraje", vehiculo.getKilometraje());
+            values.put("kilometraje_actual", vehiculo.getKilometrajeActual());
+            values.put("kilometraje_esperado", vehiculo.getKilometrajeEsperado());
 
             return db.insert(table, null, values);
         }catch (Exception e){
@@ -66,15 +72,33 @@ public class ModeloVehiculo {
     public int actualizar(ModeloVehiculo vehiculo) {
 //        try(SQLiteDatabase db = dbHelper.getWritableDatabase()){
         try {
+            if (db == null) {
+                throw new Exception("Error al abrir la base de datos");
+            }
             ContentValues values = new ContentValues();
             values.put("marca", vehiculo.getMarca());
             values.put("anho", vehiculo.getAño());
             values.put("placa", vehiculo.getPlaca());
             values.put("tipo", vehiculo.getTipo());
-            values.put("kilometraje", vehiculo.getKilometraje());
+            values.put("kilometraje_actual", vehiculo.getKilometrajeActual());
+            values.put("kilometraje_esperado", vehiculo.getKilometrajeEsperado());
 
             return db.update(table, values, "id = ?",
                     new String[]{String.valueOf(vehiculo.getId())});
+        }catch (Exception e){
+            throw new RuntimeException(e);
+        }
+    }
+    // Actualizar kilometrajes
+    public int actualizarKilometrajes(int id, int kilometraje_actual, int kilometraje_esperado) {
+        try {
+            if (db == null) {
+                throw new Exception("Error al abrir la base de datos");
+            }
+            ContentValues values = new ContentValues();
+            values.put("kilometraje_actual", kilometraje_actual);
+            values.put("kilometraje_esperado", kilometraje_esperado);
+            return db.update(table, values, "id = ?", new String[]{String.valueOf(id)});
         }catch (Exception e){
             throw new RuntimeException(e);
         }
@@ -84,6 +108,9 @@ public class ModeloVehiculo {
     public List<ModeloVehiculo> obtenerTodos() {
 //        try(SQLiteDatabase db = dbHelper.getWritableDatabase()){
         try {
+            if (db == null) {
+                throw new Exception("Error al abrir la base de datos");
+            }
             List<ModeloVehiculo> vehiculos = new ArrayList<>();
             Cursor cursor = db.query(table, null, null, null, null, null, null);
 
@@ -95,7 +122,8 @@ public class ModeloVehiculo {
                     vehiculo.setAño(cursor.getString(cursor.getColumnIndex("anho")));
                     vehiculo.setPlaca(cursor.getString(cursor.getColumnIndex("placa")));
                     vehiculo.setTipo(cursor.getString(cursor.getColumnIndex("tipo")));
-                    vehiculo.setKilometraje(cursor.getInt(cursor.getColumnIndex("kilometraje")));
+                    vehiculo.setKilometrajeActual(cursor.getInt(cursor.getColumnIndex("kilometraje_actual")));
+                    vehiculo.setKilometrajeEsperado(cursor.getInt(cursor.getColumnIndex("kilometraje_esperado")));
                     vehiculos.add(vehiculo);
                 } while (cursor.moveToNext());
             }
@@ -110,6 +138,9 @@ public class ModeloVehiculo {
     public ModeloVehiculo obtenerPorId(int id) {
 //        try (SQLiteDatabase db = dbHelper.getWritableDatabase()){
         try {
+            if (db == null) {
+                throw new Exception("Error al abrir la base de datos");
+            }
             Cursor cursor = db.query(table, null, "id = ?",
                     new String[]{String.valueOf(id)}, null, null, null);
 
@@ -120,7 +151,8 @@ public class ModeloVehiculo {
                 vehiculo.setAño(cursor.getString(cursor.getColumnIndex("anho")));
                 vehiculo.setPlaca(cursor.getString(cursor.getColumnIndex("placa")));
                 vehiculo.setTipo(cursor.getString(cursor.getColumnIndex("tipo")));
-                vehiculo.setKilometraje(cursor.getInt(cursor.getColumnIndex("kilometraje")));
+                vehiculo.setKilometrajeActual(cursor.getInt(cursor.getColumnIndex("kilometraje_actual")));
+                vehiculo.setKilometrajeEsperado(cursor.getInt(cursor.getColumnIndex("kilometraje_esperado")));
                 cursor.close();
                 return vehiculo;
             }
@@ -133,9 +165,32 @@ public class ModeloVehiculo {
     public int eliminar(int id) {
 //        try(SQLiteDatabase db = dbHelper.getWritableDatabase()){
         try {
+            if (db == null) {
+                throw new Exception("Error al abrir la base de datos");
+            }
             return db.delete(table, "id = ?",
                     new String[]{String.valueOf(id)});
         }catch (Exception e){
+            throw new RuntimeException(e);
+        }
+    }
+    @SuppressLint("Range")
+    public boolean necesitaMantenimiento(int id) {
+        try {
+            if (db == null) {
+                throw new Exception("Error al abrir la base de datos");
+            }
+            Cursor cursor = db.query(table, null, "id = ?",
+                    new String[]{String.valueOf(id)}, null, null, null);
+
+            if (cursor != null && cursor.moveToFirst()) {
+                int kilometraje_inicial = cursor.getInt(cursor.getColumnIndex("kilometraje_actual"));
+                int kilometraje_esperado = cursor.getInt(cursor.getColumnIndex("kilometraje_esperado"));
+                cursor.close();
+                return kilometraje_inicial >= kilometraje_esperado;
+            }
+            return false;
+        } catch (Exception e) {
             throw new RuntimeException(e);
         }
     }
@@ -170,10 +225,16 @@ public class ModeloVehiculo {
     public void setTipo(String tipo) {
         this.tipo = tipo;
     }
-    public int getKilometraje() {
-        return kilometraje;
+    public int getKilometrajeActual() {
+        return kilometraje_actual;
     }
-    public void setKilometraje(int kilometraje) {
-        this.kilometraje = kilometraje;
+    public void setKilometrajeActual(int kilometraje) {
+        this.kilometraje_actual = kilometraje;
+    }
+    public int getKilometrajeEsperado() {
+        return kilometraje_esperado;
+    }
+    public void setKilometrajeEsperado(int kilometraje_esperado) {
+        this.kilometraje_esperado = kilometraje_esperado;
     }
 }
